add qnas

+++ src/k_backjoon/A1300_1.java
@@ -11,11 +11,11 @@ class A1300_1 {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int N = Integer.parseInt(br.readLine());
 		int k = Integer.parseInt(br.readLine());
-		int[][] A = new int[N+1][N+1];
+		double[][] A = new double[N+1][N+1];
 		for(int i = 1; i <= N; i++) {
 			for(int j = 1; j <= N; j++) A[i][j] = i * j;
 		}
-		int[] B = new int[k+1];
+		double[] B = new double[k+1];
 		int idx = 1;
 		out : for(int i = 1; i <= N; i++) {
 			for(int j = 1; j <= N; j++) {
